Migrate S3 to the Multi-Int-Store
