+++ ingest/build.gradle
@@ -20,7 +20,6 @@ dependencies {
     implementation(Dependencies.commonsIO())
     implementation(Dependencies.ingestAPI())
     implementation(Dependencies.transformAPI())
-    implementation(Dependencies.awsS3())
     implementation project(":retrieve-api")
     testImplementation(Dependencies.springBootStarterTest())
 }
