+++ ingest/src/main/java/com/connexta/ingest/service/impl/RetrieveServiceImpl.java
@@ -6,24 +6,26 @@
  */
 package com.connexta.ingest.service.impl;
 
-import com.connexta.ingest.adaptors.S3StorageAdaptor;
 import com.connexta.ingest.service.api.RetrieveResponse;
 import com.connexta.ingest.service.api.RetrieveService;
 import java.io.IOException;
-import javax.validation.constraints.NotNull;
+import org.springframework.core.io.ByteArrayResource;
+import org.springframework.http.MediaType;
 import org.springframework.stereotype.Service;
 
 @Service
 public class RetrieveServiceImpl implements RetrieveService {
 
-  @NotNull private final S3StorageAdaptor s3Adaptor;
+  // @NotNull private final S3StorageAdaptor s3Adaptor;
 
-  public RetrieveServiceImpl(@NotNull final S3StorageAdaptor s3Adaptor) {
-    this.s3Adaptor = s3Adaptor;
-  }
+  //  public RetrieveServiceImpl(@NotNull final S3StorageAdaptor s3Adaptor) {
+  //    this.s3Adaptor = s3Adaptor;
+  //  }
 
   @Override
   public RetrieveResponse retrieve(final String ingestId) throws IOException {
-    return s3Adaptor.retrieve(ingestId);
+    // return s3Adaptor.retrieve(ingestId);
+    return new RetrieveResponse(
+        MediaType.APPLICATION_OCTET_STREAM, new ByteArrayResource("asdf".getBytes()));
   }
 }
