+++ ingest/src/test/resources/application.yml
@@ -2,5 +2,4 @@
 endpointUrl:
   transform: https://localhost/transform
   ingest:
-    callback: https://change.me
     retrieve: https://change.me
