+++ multi-int-store/src/main/java/com/connexta/multiintstore/common/exceptions/StorageException.java
@@ -4,7 +4,7 @@
  * Released under the GNU Lesser General Public License version 3; see
  * https://www.gnu.org/licenses/lgpl-3.0.html
  */
-package com.connexta.ingest.exceptions;
+package com.connexta.multiintstore.common.exceptions;
 
 public class StorageException extends Exception {
   /**
