+++ multi-int-store/src/test/java/resources/test.yml
@@ -1,0 +1,7 @@
+callback-accept-version: 0.1.0-SNAPSHOT
+aws:
+  s3:
+    endpointUrl: https://change.me
+    region: us-west-1
+    bucket:
+      quarantine: test-bucket
